Fix windows 2

+++ src/ftxui/component/terminal_input_parser.cpp
@@ -27,7 +27,7 @@ unsigned char TerminalInputParser::Current() {
 
 bool TerminalInputParser::Eat() {
   position_++;
-  return position_ < pending_.size();
+  return position_ < (int)pending_.size();
 }
 
 void TerminalInputParser::Send(TerminalInputParser::Type type) {
